property: move complete to a separate module
Until the minimum version supported by the main module is Go 1.18 which
supports generics.
