+++ .circleci/config.yml
@@ -17,11 +17,17 @@ workflows:
           executor:
             name: go/golang
             tag:  1.18-alpine
+          post-steps: &xgenerics
+            - run:
+                name: "test x/generics"
+                working_directory: ./x/generics
+                command: gotestsum -ftestname
       - go/test:
           name: test-golang-1.19
           executor:
             name: go/golang
             tag:  1.19-alpine
+          post-steps: *xgenerics
       - go/test:
           name: test-windows
           executor: windows
@@ -55,4 +61,8 @@ jobs:
     - go/install: {package: git}
     - run:
         name: Lint
+        command: golangci-lint run -v --concurrency 2
+    - run:
+        name: Lint x/generics
+        working_directory: ./x/generics
         command: golangci-lint run -v --concurrency 2
