+++ x/generics/go.mod
@@ -1,0 +1,7 @@
+module gotest.tools/x/generics
+
+go 1.18
+
+require gotest.tools/v3 v3.3.0
+
+require github.com/google/go-cmp v0.5.5 // indirect
