+++ x/generics/property/doc.go
@@ -1,0 +1,5 @@
+/*
+Package property provides experimental property assertions for writing
+property based tests.
+*/
+package property
